Training 20230411

+++ deep_events/gaussians_to_training.py
@@ -9,22 +9,32 @@ from bson.objectid import ObjectId
 import numpy as np
 from multiprocessing import Pool
 
-folder = "//lebsrv2.epfl.ch/LEB_SHARED/SHARED/_Lab members/Emily/"
-db_files = list(Path(folder).rglob(r'db.yaml'))
+folder = Path("//lebsrv2.epfl.ch/LEB_SHARED/SHARED/_Lab members/Emily/")
 SAVING_SCHEME = "ws_0.2"
 
 
 def extract_events(db_file, images_identifier: str = "", channel_contrast: str = None):
 
     tif_identifier = r'*' + images_identifier + r'*.ome.tif'
-    tif_file = sorted(Path(os.path.dirname(db_file)).glob(tif_identifier), key=os.path.getmtime)[-1]
+    tif_files = sorted(Path(os.path.dirname(db_file)).glob(tif_identifier), key=os.path.getmtime)
+    if tif_files:
+        tif_file = tif_files[-1]
+        print(tif_file)
+    else:
+        print(db_file)
+        print("Did not find a corresponding tif file for this db_file")
+        return
+
 
     if os.path.exists(os.path.join(os.path.dirname(db_file), 'ground_truth.tiff')):
         gaussians_file = os.path.join(os.path.dirname(db_file), 'ground_truth.tiff')
     elif  os.path.exists(os.path.join(os.path.dirname(db_file), 'ground_truth.tif')):
         gaussians_file = os.path.join(os.path.dirname(db_file), 'ground_truth.tif')
     else:
+        print(db_file)
+        print("Did not find a corresponding ground truth file for this db_file")
         return
+    print(gaussians_file)
 
     folder_dict = get_dict(Path(os.path.dirname(db_file)))
     event_dict = copy.deepcopy(folder_dict)
@@ -44,9 +54,9 @@ def extract_events(db_file, images_identifier: str = "", channel_contrast: str =
             event_dict = handle_db(event, box, event_dict)
 
             tifffile.imwrite(os.path.join(event_dict['event_path'], "ground_truth.tif"),
-                             (gaussians_crop).astype(np.uint16), photometric='minisblack')
+                             (gaussians_crop).astype(np.float16), photometric='minisblack')
             tifffile.imwrite(os.path.join(event_dict['event_path'], "images.tif"),
-                             (imgs_crop).astype(np.uint16), photometric='minisblack')
+                             (imgs_crop).astype(np.float16), photometric='minisblack')
 
 
 def handle_db(event, box, event_dict):
@@ -65,6 +75,8 @@ def handle_db(event, box, event_dict):
 
 
 if __name__ == "__main__":
-    shutil.rmtree(os.path.join(folder, "event_data"))
-    with Pool(12) as p:
+    if (folder / "event_data").is_dir():
+        shutil.rmtree(os.path.join(folder, "event_data"))
+    db_files = list((folder).rglob(r'db.yaml'))
+    with Pool(10) as p:
         p.starmap(extract_events, zip(db_files, ["GFP"]*len(db_files)))
